Move UpNext and StillWatching dialog class to new dialog.py
